feat: automatically infer if resources need extract
