+++ src/pymovements/gaze/transforms.py
@@ -29,7 +29,7 @@ import numpy as np
 import polars as pl
 import scipy
 
-from pymovements.utils import checks
+from pymovements._utils import _checks
 
 TransformMethod = TypeVar('TransformMethod', bound=Callable[..., pl.Expr])
 
@@ -183,8 +183,8 @@ def downsample(
     pl.Expr
         The respective polars expression.
     """
-    checks.check_is_int(factor=factor)
-    checks.check_is_positive_value(factor=factor)
+    _checks.check_is_int(factor=factor)
+    _checks.check_is_positive_value(factor=factor)
 
     return pl.all().gather_every(n=factor)
 
@@ -381,8 +381,8 @@ def _check_distance(distance: float) -> None:
     distance: float
         The distance to check.
     """
-    checks.check_is_scalar(distance=distance)
-    checks.check_is_greater_than_zero(distance=distance)
+    _checks.check_is_scalar(distance=distance)
+    _checks.check_is_greater_than_zero(distance=distance)
 
 
 def _check_screen_resolution(screen_resolution: tuple[int, int]) -> None:
@@ -408,8 +408,8 @@ def _check_screen_resolution(screen_resolution: tuple[int, int]) -> None:
         )
 
     for element in screen_resolution:
-        checks.check_is_scalar(screen_resolution=element)
-        checks.check_is_greater_than_zero(screen_resolution=element)
+        _checks.check_is_scalar(screen_resolution=element)
+        _checks.check_is_greater_than_zero(screen_resolution=element)
 
 
 def _check_screen_size(screen_size: tuple[float, float]) -> None:
@@ -433,8 +433,8 @@ def _check_screen_size(screen_size: tuple[float, float]) -> None:
         raise ValueError(f'screen_size must have length of 2, but is of length {len(screen_size)}')
 
     for element in screen_size:
-        checks.check_is_scalar(screen_size=element)
-        checks.check_is_greater_than_zero(screen_size=element)
+        _checks.check_is_scalar(screen_size=element)
+        _checks.check_is_greater_than_zero(screen_size=element)
 
 
 @register_transform
@@ -773,7 +773,7 @@ def resample(
     elif isinstance(columns, str):
         columns = [columns]
 
-    checks.check_is_greater_than_zero(resampling_rate=resampling_rate)
+    _checks.check_is_greater_than_zero(resampling_rate=resampling_rate)
 
     # Return frame if empty
     if frame.is_empty():
@@ -1134,9 +1134,9 @@ def _check_window_length(window_length: Any) -> None:
     window_length: Any
         The window length to check.
     """
-    checks.check_is_not_none(window_length=window_length)
-    checks.check_is_int(window_length=window_length)
-    checks.check_is_greater_than_zero(degree=window_length)
+    _checks.check_is_not_none(window_length=window_length)
+    _checks.check_is_int(window_length=window_length)
+    _checks.check_is_greater_than_zero(degree=window_length)
 
 
 def _check_degree(degree: Any, window_length: int) -> None:
@@ -1150,9 +1150,9 @@ def _check_degree(degree: Any, window_length: int) -> None:
     window_length: int
         The window length to check against.
     """
-    checks.check_is_not_none(degree=degree)
-    checks.check_is_int(degree=degree)
-    checks.check_is_greater_than_zero(degree=degree)
+    _checks.check_is_not_none(degree=degree)
+    _checks.check_is_int(degree=degree)
+    _checks.check_is_greater_than_zero(degree=degree)
 
     if degree >= window_length:
         raise ValueError("'degree' must be less than 'window_length'")
@@ -1191,5 +1191,5 @@ def _check_derivative(derivative: Any) -> None:
     derivative: Any
         The derivative to check.
     """
-    checks.check_is_int(derivative=derivative)
-    checks.check_is_positive_value(derivative=derivative)
+    _checks.check_is_int(derivative=derivative)
+    _checks.check_is_positive_value(derivative=derivative)
