+++ src/pymovements/plotting/scanpathplot.py
@@ -31,9 +31,9 @@ from matplotlib.patches import Circle
 
 from pymovements.events import EventDataFrame
 from pymovements.gaze import GazeDataFrame
-from pymovements.utils.plotting import draw_line_data
-from pymovements.utils.plotting import LinearSegmentedColormapType
-from pymovements.utils.plotting import setup_matplotlib
+from pymovements.plotting._matlab import draw_line_data
+from pymovements.plotting._matlab import LinearSegmentedColormapType
+from pymovements.plotting._matlab import setup_matplotlib
 
 
 # This is really a dirty workaround to use the Agg backend if runnning pytest.
